+++ custom_components/havcs/const.py
@@ -2,3 +2,6 @@
 CONF_BROKER = 'broker'
 CONF_DISCOVERY = 'discovery'
 DEFAULT_DISCOVERY = False
+INTEGRATION = 'havcs'
+STORAGE_VERSION = 1
+STORAGE_KEY = 'havcs'
