+++ custom_components/havcs/services.yaml
@@ -1,0 +1,2 @@
+reload:
+  description: 重新加载设备配置。
